+++ admin/src/components/UserList.jsx
@@ -6,6 +6,7 @@ import Button from "./ui/Button";
 import ButtonAction from "./ui/ButtonAction";
 import { RiApps2AddFill } from "react-icons/ri";
 import { MdEditSquare, MdDelete, MdSearch } from "react-icons/md";
+import { GrPowerReset } from "react-icons/gr";
 import { BiSolidUserDetail, BiStats } from "react-icons/bi";
 import ReactPaginate from "react-paginate";
 
@@ -67,24 +68,57 @@ const UserList = () => {
     }
   };
 
-  const deleteUser = async (userId) => {
-    await axios.delete(`${API_URL}/users/user/${userId}`);
-    getUsers();
-
+  const handleReset = async (id) => {
     Swal.fire({
-      icon: "success",
-      title: "Success",
-      text: "User deleted successfully",
+      title: "Apakah Anda yakin?",
+      text: "Password akan direset ke default!",
+      icon: "warning",
+      showCancelButton: true,
+      confirmButtonText: "Ya, reset!",
+    }).then(async (result) => {
+      if (result.isConfirmed) {
+        try {
+          await axios.put(`${API_URL}/auth-web/update-pass/${id}`);
+          Swal.fire("Berhasil!", "Password telah direset.", "success");
+        } catch (error) {
+          Swal.fire("Error!", error.response?.data?.message || "Terjadi kesalahan", "error");
+        }
+      }
     });
   };
+  
+
+  const deleteUser = async (userId) => {
+    Swal.fire({
+      title: "Are you sure?",
+      text: "You won't be able to revert this!",
+      icon: "warning",
+      showCancelButton: true,
+      confirmButtonColor: "#d33",
+      cancelButtonColor: "#3085d6",
+      confirmButtonText: "Yes, delete it!",
+    }).then(async (result) => {
+      if (result.isConfirmed) {
+        await axios.delete(`${API_URL}/users/user/${userId}`);
+        getUsers();
+  
+        Swal.fire({
+          icon: "success",
+          title: "Deleted!",
+          text: "User deleted successfully.",
+        });
+      }
+    });
+  };
+  
 
   return (
     <>
       <div>
         <h2 className="text-2xl font-semibold mb-4">Users</h2>
 
-        <div className="flex justify-between items-center">
-        <Button
+        <div className="flex gap-2 justify-between items-center">
+          <Button
             text="Add New"
             to="/users/add"
             iconPosition="left"
@@ -128,8 +162,6 @@ const UserList = () => {
               </div>
             </form>
           </div>
-
-         
         </div>
 
         <div className="mt-5 overflow-x-auto bg-white rounded-xl p-4">
@@ -151,7 +183,22 @@ const UserList = () => {
                     <td className="px-4 py-2 border-b">{index + 1}</td>
                     <td className="px-4 py-2 border-b">{user.username}</td>
                     <td className="px-4 py-2 border-b">{user.email}</td>
-                    <td className="px-4 py-2 border-b">{user.role}</td>
+                    <td className="px-4 py-2 border-b">
+                      {user.role === "admin" ? (
+                        <span className="px-2 py-1 text-xs text-orange-600 border border-orange-600 rounded-lg">
+                          Admin
+                        </span>
+                      ) : user.role === "user" ? (
+                        <span className="px-2 py-1 text-xs text-green-600 border border-green-600 rounded-lg">
+                          User
+                        </span>
+                      ) : (
+                        <span className="px-2 py-1 text-xs text-blue-600 border border-blue-600 rounded-lg">
+                          Delivery
+                        </span>
+                      )}
+                    </td>
+
                     <td className="px-4 py-2 border-b">
                       <div className="flex gap-x-2">
                         <ButtonAction
@@ -168,6 +215,11 @@ const UserList = () => {
                           to={`/users/${user.id}/stats`}
                           icon={<BiStats />}
                           className={"bg-purple-600 hover:bg-purple-700"}
+                        />
+                        <ButtonAction
+                          onClick={() => handleReset(user.id)}
+                          icon={<GrPowerReset />}
+                          className={"bg-green-600 hover:bg-green-700"}
                         />
                         <ButtonAction
                           onClick={() => deleteUser(user.id)}
